requirements.txt
Flask==2.1.0
google-generativeai==0.3.1